+++ cli/lib/cmd_hardware.sh
@@ -31,7 +31,7 @@ cmd_read() {
 
             local kind_color="c_sensor"
             case "$dkind" in
-                digital_out|relay|pwm|rgb_led) kind_color="c_actuator" ;;
+                digital_out|relay|pwm|rgb_led|ssd1306|sh1106) kind_color="c_actuator" ;;
             esac
 
             printf '\n'
@@ -511,7 +511,7 @@ cmd_devices() {
 
         local kind_color="c_sensor"
         case "$dkind" in
-            digital_out|relay|pwm|rgb_led) kind_color="c_actuator" ;;
+            digital_out|relay|pwm|rgb_led|ssd1306) kind_color="c_actuator" ;;
         esac
 
         printf '  %s%-16s%s %s%-14s%s %s%-10s%s %s%s%s\n' \